myndla-api: move robot related business logic to `RobotService`
