+++ myndla-api/src/main/scala/no/ndla/myndlaapi/service/FolderWriteService.scala
@@ -17,7 +17,6 @@ import no.ndla.common.model.domain.ResourceType
 import no.ndla.common.model.domain.myndla.{FolderStatus, MyNDLAUser}
 import no.ndla.database.DBUtility
 import no.ndla.myndlaapi.integration.SearchApiClient
-import no.ndla.myndlaapi.model.api.robot.{CreateRobotDefinitionDTO, RobotDefinitionDTO}
 import no.ndla.myndlaapi.model.domain.FolderSortObject.{
   FolderSorting,
   ResourceSorting,
@@ -40,9 +39,6 @@ import no.ndla.myndlaapi.model.domain.{
   FolderAndDirectChildren,
   FolderSortException,
   Rankable,
-  RobotConfiguration,
-  RobotDefinition,
-  RobotStatus,
   SavedSharedFolder
 }
 import no.ndla.myndlaapi.repository.{FolderRepository, RobotRepository, UserRepository}
@@ -85,61 +81,6 @@ trait FolderWriteService {
           }
       })
     }
-
-    def updateRobot(
-        robotId: UUID,
-        robotDefinitionDTO: CreateRobotDefinitionDTO,
-        feideToken: Option[FeideAccessToken]
-    ): Try[RobotDefinitionDTO] = updateRobotWith(robotId, feideToken) {
-      _.copy(
-        status = robotDefinitionDTO.status,
-        configuration = RobotConfiguration.fromDTO(robotDefinitionDTO.configuration)
-      )
-    }
-
-    def updateRobotStatus(
-        robotId: UUID,
-        newStatus: RobotStatus,
-        feideToken: Option[FeideAccessToken]
-    ): Try[RobotDefinitionDTO] =
-      updateRobotWith(robotId, feideToken) {
-        _.copy(status = newStatus)
-      }
-
-    def deleteRobot(robotId: UUID, feideToken: Option[FeideAccessToken]): Try[Unit] = DBUtil.rollbackOnFailure {
-      session =>
-        for {
-          feideId       <- feideApiClient.getFeideID(feideToken)
-          _             <- canWriteDuringMyNDLAWriteRestrictionsOrAccessDenied(feideId, feideToken)
-          maybeRobot    <- robotRepository.getRobotWithId(robotId)(session)
-          existingRobot <- maybeRobot.toTry(NotFoundException(s"Could not find editable robot with id '$robotId'"))
-          _             <- existingRobot.canEdit(feideId)
-          _             <- robotRepository.deleteRobotDefinition(robotId)(session)
-        } yield ()
-    }
-
-    private def updateRobotWith(robotId: UUID, feideToken: Option[FeideAccessToken])(
-        updateFunc: RobotDefinition => RobotDefinition
-    ): Try[RobotDefinitionDTO] =
-      DBUtil.rollbackOnFailure { session =>
-        val now = clock.now()
-        for {
-          feideId       <- feideApiClient.getFeideID(feideToken)
-          _             <- canWriteDuringMyNDLAWriteRestrictionsOrAccessDenied(feideId, feideToken)
-          maybeRobot    <- robotRepository.getRobotWithId(robotId)(session)
-          existingRobot <- maybeRobot.toTry(NotFoundException(s"Could not find editable robot with id '$robotId'"))
-          _             <- existingRobot.canEdit(feideId)
-          updated       <- Try(updateFunc(existingRobot))
-          sharedTime = updated.status match {
-            case RobotStatus.SHARED if existingRobot.shared.isEmpty => Some(now)
-            case RobotStatus.SHARED                                 => existingRobot.shared
-            case _                                                  => None
-          }
-          withUpdatedTimes = updated.copy(updated = now, shared = sharedTime)
-          _ <- robotRepository.updateRobotDefinition(withUpdatedTimes)(session)
-
-        } yield RobotDefinitionDTO.fromDomain(withUpdatedTimes)
-      }
 
     private def canWriteNow(myNDLAUser: MyNDLAUser): Try[Boolean] = {
       if (myNDLAUser.isTeacher) return Success(true)
@@ -801,26 +742,5 @@ trait FolderWriteService {
         })
     }
 
-    def createRobot(
-        robotDefinitionDTO: CreateRobotDefinitionDTO,
-        feideToken: Option[FeideAccessToken]
-    ): Try[RobotDefinitionDTO] =
-      DBUtil.rollbackOnFailure { session =>
-        val now = clock.now()
-        for {
-          feideId <- feideApiClient.getFeideID(feideToken)
-          _       <- canWriteDuringMyNDLAWriteRestrictionsOrAccessDenied(feideId, feideToken)
-          domain = RobotDefinition(
-            id = UUID.randomUUID(),
-            feideId = feideId,
-            status = robotDefinitionDTO.status,
-            configuration = RobotConfiguration.fromDTO(robotDefinitionDTO.configuration),
-            created = now,
-            updated = now,
-            shared = Option.when(robotDefinitionDTO.status == RobotStatus.SHARED)(now)
-          )
-          robot <- robotRepository.insertRobotDefinition(domain)(session)
-        } yield RobotDefinitionDTO.fromDomain(robot)
-      }
   }
 }
